requirements.txt
# ClearCut Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
python-dotenv>=1.0.0
